backend/requirements.txt
fastapi
uvicorn
SQLAlchemy
email-validator
psycopg2-binary
python-dotenv
pydantic-extra-types
pydantic-settings
python-jose
bcrypt==3.2.2
passlib
python-multipart